+++ recipes/install.rb
@@ -22,7 +22,7 @@ install_type = node['kibana']['install_type']
 unless Chef::Config[:solo]
   es_server_results = search(:node, "roles:#{node['kibana']['es_role']} AND chef_environment:#{node.chef_environment}")
   unless es_server_results.empty?
-    node.set['kibana']['es_server'] = es_server_results[0]['ipaddress']
+    node.override['kibana']['es_server'] = es_server_results[0]['ipaddress']
   end
 end
 
@@ -54,24 +54,15 @@ end
 
 docroot = "#{node['kibana']['install_dir']}/current/kibana"
 kibana_config = "#{node['kibana']['install_dir']}/current/#{node['kibana'][install_type]['config']}"
-es_server = "#{node['kibana']['es_scheme']}#{node['kibana']['es_server']}:#{node['kibana']['es_port']}"
 
-config_template = Gem::Version.new(node['kibana']['version'].split('-')[0]) < Gem::Version.new('6.0.0')?'config_template_5':'config_template'
+node.override['kibana']['config']['elasticsearch.url'] = "#{node['kibana']['es_scheme']}#{node['kibana']['es_server']}:#{node['kibana']['es_port']}" unless node['kibana']['config']['elasticsearch.url']
 
-template kibana_config do
-  source node['kibana'][install_type][config_template]
-  cookbook node['kibana'][install_type]['config_template_cookbook']
-  mode '0644'
-  user kibana_user
+# generate yml kibana configuration from json (converted)
+file ::File.join(kibana_config) do
+  content YAML.dump(node['kibana']['config'].to_hash)
+  owner kibana_user
   group kibana_user
-  variables(
-    index: node['kibana']['config']['kibana_index'],
-    port: node['kibana']['java_webserver_port'],
-    elasticsearch: es_server,
-    default_route: node['kibana']['config']['default_route'],
-    panel_names:  node['kibana']['config']['panel_names'],
-    default_app_id: node['kibana']['config']['default_app_id']
-  )
+  mode '0644'
 end
 
 if install_type == 'file'
@@ -99,7 +90,7 @@ kibana_web 'kibana' do
   type lazy { node['kibana']['webserver'] }
   docroot docroot
   es_server node['kibana']['es_server']
-  kibana_port node['kibana']['java_webserver_port']
+  kibana_port node['kibana']['config']['server.port']
   template node['kibana'][webserver]['template']
   template_cookbook node['kibana'][webserver]['template_cookbook']
   not_if { node['kibana']['webserver'] == '' }